scrape/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
pandas==2.2.1
ratelimit==2.2.1